requirements.txt
Flask==2.3.3
tensorflow==2.15.0
numpy==1.24.4
opencv-python==4.9.0.80
Pillow==10.3.0
scikit-learn==1.3.2
matplotlib==3.8.4
